+++ modules/platforms/end4-pC/tempconfig.lua
@@ -1,0 +1,27 @@
+hl.monitor({
+    output = "DP-3",
+    mode = "2560x1440@179.959Hz",
+    position = "0x0",
+    scale = 1
+})
+
+hl.monitor({
+    output = "DP-2",
+    mode = "1920x1080@74.973Hz",
+    position = "2560x0",
+    scale = 1
+})
+
+hl.monitor({
+    output = "HDMI-A-1",
+    mode = "1920x1080@143.981Hz",
+    position = "320x-1080",
+    scale = 1
+})
+
+hl.config ({
+    input = {
+        kb_layout = "gb",
+        kb_variant = ""
+    }
+})