+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 #### Changed
 - Replaced Nix IDE with bbenoist's nix within the `vscodium` homeModule.
 
+#### Fixed
+- Fixed the `end4-pC` input reverting to the original hyprland config via `tempconfig`.
+
 
 ## [v0.4.0]